ead_api/requirements.txt
bcrypt==4.0.1
Django==4.2.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
Pillow==9.2.0
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2022.5
